Recommand Etalab app on DVF datasets
I've noticed that on most case a discussion is needed to choose which reuse to recommend. Also it is frequent that the reuses are not referenced (https://annuaire-entreprises.data.gouv.fr/ for the Siren/Siret datasets for instance). I'll try to think of something to increase the number of reuses referenced.